+++ src/runtime_src/core/common/xclbin_parser.h
@@ -62,7 +62,7 @@ struct kernel_properties
   mailbox_type mailbox = mailbox_type::none;
   size_t address_range = 0x10000;  // NOLINT, default address range
   bool sw_reset = false;
-  uint32_t functional = 0;
+  size_t functional = 0;
 
   // opencl specifics
   size_t workgroupsize = 0;
